Fixed failed to load ApplicationContext in @SpringBootTest
The application config for tests was corrupted and due to that Spring could not load ApplicationContext when performing a SpringBootTest.
The Keycloak Config section is marked as TODO because as of now it is not clear whether the config is really necessary for testing. Will be removed in future commit